Speeding up compile/link time for applications using partitioned_vector
- by pre-compiling partitioned_vector<double>, partitioned_vector<std::string>,
  and partitioned_vector<int> into partitioned_vector component library
- fixing all tests to use the pre-compiled implementations
