+++ hpx/components/containers/coarray/coarray.hpp
@@ -25,15 +25,19 @@
 ///////////////////////////////////////////////////////////////////////////////
 /// \cond NOINTERNAL
 
+#define HPX_REGISTER_COARRAY_DECLARATION(...)                                  \
+    HPX_REGISTER_PARTITIONED_VECTOR_DECLARATION(__VA_ARGS__)
 #define HPX_REGISTER_COARRAY(...) HPX_REGISTER_PARTITIONED_VECTOR(__VA_ARGS__)
 
 namespace hpx
 {
-    namespace detail{
-
+    namespace detail
+    {
         struct auto_subscript
         {
-            constexpr auto_subscript(){}
+            constexpr auto_subscript()
+            {
+            }
 
             // Defined to pass the coarray_sizes constructor
             operator std::size_t()
@@ -42,50 +46,48 @@ namespace hpx
             }
         };
 
-        template<typename C>
+        template <typename C>
         struct cast_if_autosubscript
         {
             using type = C;
         };
 
-        template<>
+        template <>
         struct cast_if_autosubscript<detail::auto_subscript>
         {
             using type = int;
         };
 
-        template<std::size_t N>
+        template <std::size_t N>
         struct coarray_sizes
         {
-            using iterator =
-                typename std::array<std::size_t,N>::iterator;
+            using iterator = typename std::array<std::size_t, N>::iterator;
 
             using const_iterator =
-                typename std::array<std::size_t,N>::const_iterator;
+                typename std::array<std::size_t, N>::const_iterator;
 
-            template<typename ... I>
+            template <typename... I>
             coarray_sizes(I... i)
-            : sizes_({{i...}})
+              : sizes_({{i...}})
             {
                 using last_element =
-                    typename util::detail::at_index<sizeof...(I)-1, I...>::type;
+                    typename util::detail::at_index<sizeof...(I) - 1, I...>::type;
                 using condition =
-                    typename std::is_same<last_element,detail::auto_subscript>;
+                    typename std::is_same<last_element, detail::auto_subscript>;
 
                 static_assert(condition::value,
-                    "hpx::coarray() needs the last size to be " \
-                    "equal to hpx::_");
+                    "hpx::coarray() needs the last size to be equal to hpx::_");
 
                 static_assert(N == sizeof...(I),
-                    "hpx::coarray() needs the number of sizes to be " \
+                    "hpx::coarray() needs the number of sizes to be "
                     "equal to its dimension");
 
-                static_assert( util::detail::all_of<
-                    typename std::is_integral<
-                        typename detail::cast_if_autosubscript<I>::type
-                    >::type ... >::value,
-                        "One or more elements in sizes given to hpx::coarray() "\
-                        "is not integral");
+                static_assert(
+                    util::detail::all_of<typename std::is_integral<
+                        typename detail::cast_if_autosubscript<I>::type>::
+                            type...>::value,
+                    "One or more elements in sizes given to hpx::coarray() "
+                    "is not integral");
             }
 
             iterator begin()
@@ -109,80 +111,76 @@ namespace hpx
             }
 
         private:
-            std::array<std::size_t,N> sizes_;
+            std::array<std::size_t, N> sizes_;
         };
-
     }
 
     // Used for "automatic" coarray subscript and "automatic" coarray size
-    namespace container{ namespace placeholders{
-        constexpr hpx::detail::auto_subscript _;
+    namespace container { namespace placeholders
+    {
+            constexpr hpx::detail::auto_subscript _;
     }}
 
-    template<typename T, std::size_t N, typename Data = std::vector<T>>
-    struct coarray
-    : public partitioned_vector_view<T,N,Data>
+    template <typename T, std::size_t N, typename Data = std::vector<T>>
+    struct coarray : public partitioned_vector_view<T, N, Data>
     {
     private:
-        using base_type = partitioned_vector_view<T,N,Data>;
-        using indices =
-            typename hpx::util::detail::make_index_pack<N>::type;
+        using base_type = partitioned_vector_view<T, N, Data>;
+        using indices = typename hpx::util::detail::make_index_pack<N>::type;
 
-        std::vector<hpx::naming::id_type>
-        get_unrolled_localities(
-            std::vector< hpx::naming::id_type > const & in,
+        std::vector<hpx::naming::id_type> get_unrolled_localities(
+            std::vector<hpx::naming::id_type> const& in,
             std::size_t num_segments,
             std::size_t unroll)
         {
-            using iterator = typename
-                    std::vector<hpx::naming::id_type>::iterator;
+            using iterator = typename std::vector<hpx::naming::id_type>::iterator;
 
-            using const_iterator = typename
-                    std::vector<hpx::naming::id_type>::const_iterator;
+            using const_iterator =
+                typename std::vector<hpx::naming::id_type>::const_iterator;
 
             std::vector<hpx::naming::id_type> out(num_segments);
 
             iterator o_end = out.end();
             const_iterator i_begin = in.cbegin();
-            const_iterator i_end   = in.cend();
+            const_iterator i_end = in.cend();
             const_iterator i = i_begin;
 
-            for(iterator o = out.begin(); o<o_end; o+=unroll)
+            for (iterator o = out.begin(); o < o_end; o += unroll)
             {
-               std::fill(o, (std::min)(o + unroll, o_end), *i);
-               i = (++i != i_end) ? i : i_begin;
-           }
+                std::fill(o, (std::min)(o + unroll, o_end), *i);
+                i = (++i != i_end) ? i : i_begin;
+            }
 
-           return out;
-       }
+            return out;
+        }
 
-        template<typename Iterator, std::size_t ... I>
-        base_type update_view(
-            hpx::detail::coarray_sizes<N> const & cosizes,
+        template <typename Iterator, std::size_t... I>
+        base_type update_view(hpx::detail::coarray_sizes<N> const& cosizes,
             std::size_t num_images,
             hpx::util::detail::pack_c<std::size_t, I...>,
-            hpx::lcos::spmd_block const & block,
-            Iterator && begin,
-            Iterator && last) const
+            hpx::lcos::spmd_block const& block,
+            Iterator&& begin,
+            Iterator&& last) const
         {
-            return base_type(
-                block,
+            return base_type(block,
                 std::forward<Iterator>(begin),
                 std::forward<Iterator>(last),
-                {(cosizes.begin()[I]!= std::size_t(-1)
-                    ? cosizes.begin()[I] : num_images)...});
+                {(cosizes.begin()[I] != std::size_t(-1) ? cosizes.begin()[I] :
+                                                          num_images)...});
         }
 
     public:
-        explicit coarray(
-            hpx::lcos::spmd_block const & block,
+        explicit coarray(hpx::lcos::spmd_block const& block,
             std::string name,
-            hpx::detail::coarray_sizes<N> cosizes,
+            hpx::detail::coarray_sizes<N>
+                cosizes,
             std::size_t segment_size)
-        : base_type(block), vector_(), this_image_(block.this_image())
+          : base_type(block)
+          , vector_()
+          , this_image_(block.this_image())
         {
-        // Used to access base members
-            base_type & view (*this);
+            // Used to access base members
+            base_type& view(*this);
 
             std::size_t num_images = block.get_num_images();
 
@@ -190,69 +188,55 @@ namespace hpx
             {
                 std::size_t num_segments = N > 0 ? 1 : 0;
 
-                for(std::size_t const & i : cosizes)
+                for (std::size_t const& i : cosizes)
                 {
                     num_segments *= (i != std::size_t(-1) ? i : num_images);
                 }
 
-                std::vector< hpx::id_type > localities
-                    = hpx::find_all_localities();
+                std::vector<hpx::id_type> localities = hpx::find_all_localities();
 
-                vector_ = hpx::partitioned_vector<T,Data>(
-                    segment_size*num_segments, T(0),
-                        hpx::container_layout(
-                            num_segments,
-                            get_unrolled_localities(
-                                localities, num_segments,
-                                num_segments/localities.size())
-                        )
-                    );
+                vector_ = hpx::partitioned_vector<T, Data>(
+                    segment_size * num_segments, T(0),
+                    hpx::container_layout(num_segments,
+                        get_unrolled_localities(localities, num_segments,
+                            num_segments / localities.size())));
 
                 vector_.register_as(hpx::launch::sync, name + "_hpx_coarray");
             }
             else
                 vector_.connect_to(hpx::launch::sync, name + "_hpx_coarray");
 
-            view = update_view(cosizes,
-                num_images, indices(), block, vector_.begin(), vector_.end());
+            view = update_view(cosizes, num_images, indices(), block,
+                vector_.begin(), vector_.end());
         }
 
-        template<typename... I,
-            typename = std::enable_if_t<
-                ! std::is_same<
-                    typename util::detail::at_index<sizeof...(I)-1,I...>::type,
-                        detail::auto_subscript>::value
-                >
-            >
-        hpx::detail::view_element<T,Data>
-        operator()(I... index) const
+        template <typename... I,
+            typename = std::enable_if_t<!std::is_same<
+                typename util::detail::at_index<sizeof...(I) - 1, I...>::type,
+                detail::auto_subscript>::value>>
+        hpx::detail::view_element<T, Data> operator()(I... index) const
         {
             return base_type::operator()(
-                (std::is_same<I,detail::auto_subscript>::value
-                    ? this_image_
-                    : index)...);
+                (std::is_same<I, detail::auto_subscript>::value ? this_image_ :
+                                                                  index)...);
         }
 
-        template<typename... I,
-            typename = std::enable_if_t<
-                std::is_same<
-                    typename util::detail::at_index<sizeof...(I)-1,I...>::type,
-                        detail::auto_subscript>::value
-                >
-            >
-        Data &
-        operator()(I... index)
+        template <typename... I,
+            typename = std::enable_if_t<std::is_same<
+                typename util::detail::at_index<sizeof...(I) - 1, I...>::type,
+                detail::auto_subscript>::value>>
+        Data& operator()(I... index)
         {
             return base_type::operator()(
-                (std::is_same<I,detail::auto_subscript>::value
-                    ? this_image_
-                    : index)...).data();
+                (std::is_same<I, detail::auto_subscript>::value ? this_image_ :
+                                                                  index)...)
+                .data();
         }
 
     private:
-        hpx::partitioned_vector<T,Data> vector_;
+        hpx::partitioned_vector<T, Data> vector_;
         std::size_t this_image_;
     };
 }
 
-#endif // COARRAY_HPP
+#endif    // COARRAY_HPP
