+++ hpx/include/partitioned_vector_predef.hpp
@@ -1,0 +1,13 @@
+//  Copyright (C) 2017 Hartmut Kaiser
+//
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#if !defined(HPX_PARTITIONED_VECTOR_PREDEF_AUG_02_2017_0656PM)
+#define HPX_PARTITIONED_VECTOR_PREDEF_AUG_02_2017_0656PM
+
+#include <hpx/components/containers/partitioned_vector/partitioned_vector_predef.hpp>
+
+#endif
+
+
