+++ hpx/include/partitioned_vector_view.hpp
@@ -1,0 +1,14 @@
+//  Copyright (C) 2017 Hartmut Kaiser
+//
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#if !defined(HPX_PARTITIONED_VECTOR_VIEW_AUG_03_2017_1157AM)
+#define HPX_PARTITIONED_VECTOR_VIEW_AUG_03_2017_1157AM
+
+#include <hpx/components/containers/partitioned_vector/partitioned_vector_view.hpp>
+#include <hpx/components/containers/partitioned_vector/partitioned_vector_local_view.hpp>
+
+#endif
+
+
