+++ tests/unit/component/coarray.cpp
@@ -16,8 +16,8 @@
 #include <utility>
 
 ///////////////////////////////////////////////////////////////////////////////
-// Define the coarray types to be used.
-HPX_REGISTER_COARRAY(double);
+// coarray<double> is predefined in the partitioned_vector module
+HPX_REGISTER_COARRAY_DECLARATION(double);
 
 void bulk_test( hpx::lcos::spmd_block block,
                 std::size_t height,
