+++ tests/unit/parallel/segmented_algorithms/partitioned_vector_inclusive_scan.cpp
@@ -4,8 +4,8 @@
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include <hpx/hpx_main.hpp>
-#include <hpx/include/partitioned_vector.hpp>
 #include <hpx/include/parallel_scan.hpp>
+#include <hpx/include/partitioned_vector_predef.hpp>
 
 #include <hpx/util/lightweight_test.hpp>
 
@@ -17,36 +17,34 @@
 #include <vector>
 
 ///////////////////////////////////////////////////////////////////////////////
-// Define the vector types to be used.
-HPX_REGISTER_PARTITIONED_VECTOR(double);
-HPX_REGISTER_PARTITIONED_VECTOR(int);
+// The vector types to be used are defined in partitioned_vector module.
+// HPX_REGISTER_PARTITIONED_VECTOR(double);
+// HPX_REGISTER_PARTITIONED_VECTOR(int);
 
-#define msg7(a,b,c,d,e,f,g) \
-        std::cout \
-        << std::setw(60) << a << std::setw(40) <<  b \
-        << std::setw(10) << c << std::setw(6)  << " " << #d \
-        << " " << e << " " << f << " " << g << " ";
-#define msg9(a,b,c,d,e,f,g,h,i) \
-        std::cout \
-        << std::setw(60) << a << std::setw(40) <<  b \
-        << std::setw(10) << c << std::setw(6)  << " " << #d \
-        << " " << e << " " << f << " " << g << " " << h << " " << i << " ";
+#define msg7(a, b, c, d, e, f, g)                                              \
+    std::cout << std::setw(60) << a << std::setw(40) << b << std::setw(10)     \
+              << c << std::setw(6) << " " << #d << " " << e << " " << f << " " \
+              << g << " ";
+#define msg9(a, b, c, d, e, f, g, h, i)                                        \
+    std::cout << std::setw(60) << a << std::setw(40) << b << std::setw(10)     \
+              << c << std::setw(6) << " " << #d << " " << e << " " << f << " " \
+              << g << " " << h << " " << i << " ";
 
 ///////////////////////////////////////////////////////////////////////////////
 
-template<typename T>
+template <typename T>
 struct opt
 {
-    T operator()(T v1, T v2) const{
+    T operator()(T v1, T v2) const
+    {
         return v1 + v2;
     }
 };
 
 ///////////////////////////////////////////////////////////////////////////////
 template <typename T, typename DistPolicy, typename ExPolicy>
-void inclusive_scan_algo_tests_with_policy(
-    std::size_t size, DistPolicy const& dist_policy,
-    hpx::partitioned_vector<T>& in,
+void inclusive_scan_algo_tests_with_policy(std::size_t size,
+    DistPolicy const& dist_policy, hpx::partitioned_vector<T>& in,
     std::vector<T> ver, ExPolicy const& policy)
 {
     msg7(typeid(ExPolicy).name(), typeid(DistPolicy).name(), typeid(T).name(),
@@ -60,8 +58,8 @@ void inclusive_scan_algo_tests_with_policy(
     double e1 = t1.elapsed();
     t1.restart();
 
-    hpx::parallel::inclusive_scan(policy,
-        in.begin(), in.end(), out.begin(), opt<T>(), val);
+    hpx::parallel::inclusive_scan(
+        policy, in.begin(), in.end(), out.begin(), opt<T>(), val);
 
     double e2 = t1.elapsed();
     t1.restart();
@@ -69,20 +67,21 @@ void inclusive_scan_algo_tests_with_policy(
     HPX_TEST(std::equal(out.begin(), out.end(), ver.begin()));
 
     double e3 = t1.elapsed();
-    std::cout << std::setprecision(4) << "\t" << e1 << " " << e2 << " " << e3 << "\n";
+    std::cout << std::setprecision(4) << "\t" << e1 << " " << e2 << " " << e3
+              << "\n";
 }
 
 template <typename T, typename DistPolicy, typename ExPolicy>
-void inclusive_scan_algo_tests_segmented_out_with_policy(
-    std::size_t size, DistPolicy const& in_dist_policy,
-    DistPolicy const& out_dist_policy,
+void inclusive_scan_algo_tests_segmented_out_with_policy(std::size_t size,
+    DistPolicy const& in_dist_policy, DistPolicy const& out_dist_policy,
     hpx::partitioned_vector<T>& in, hpx::partitioned_vector<T> out,
     std::vector<T> ver, ExPolicy const& policy)
 {
     msg9(typeid(ExPolicy).name(), typeid(DistPolicy).name(), typeid(T).name(),
-        segmented, size,
-        in_dist_policy.get_num_partitions(), in_dist_policy.get_localities().size(),
-        out_dist_policy.get_num_partitions(), out_dist_policy.get_localities().size());
+        segmented, size, in_dist_policy.get_num_partitions(),
+        in_dist_policy.get_localities().size(),
+        out_dist_policy.get_num_partitions(),
+        out_dist_policy.get_localities().size());
     hpx::util::high_resolution_timer t1;
 
     T val(0);
@@ -90,8 +89,8 @@ void inclusive_scan_algo_tests_segmented_out_with_policy(
     double e1 = t1.elapsed();
     t1.restart();
 
-    hpx::parallel::inclusive_scan(policy,
-        in.begin(), in.end(), out.begin(), opt<T>(), val);
+    hpx::parallel::inclusive_scan(
+        policy, in.begin(), in.end(), out.begin(), opt<T>(), val);
 
     double e2 = t1.elapsed();
     t1.restart();
@@ -99,17 +98,17 @@ void inclusive_scan_algo_tests_segmented_out_with_policy(
     verify_values(out, ver);
 
     double e3 = t1.elapsed();
-    std::cout << std::setprecision(4) << "\t" << e1 << " " << e2 << " " << e3 << "\n";
+    std::cout << std::setprecision(4) << "\t" << e1 << " " << e2 << " " << e3
+              << "\n";
 }
 
 template <typename T, typename DistPolicy, typename ExPolicy>
-void inclusive_scan_algo_tests_inplace_with_policy(
-    std::size_t size, DistPolicy const& dist_policy,
-    std::vector<T> ver, ExPolicy const& policy)
+void inclusive_scan_algo_tests_inplace_with_policy(std::size_t size,
+    DistPolicy const& dist_policy, std::vector<T> ver, ExPolicy const& policy)
 {
     msg7(typeid(ExPolicy).name(), typeid(DistPolicy).name(), typeid(T).name(),
-        inplace, size,
-        dist_policy.get_num_partitions(), dist_policy.get_localities().size());
+        inplace, size, dist_policy.get_num_partitions(),
+        dist_policy.get_localities().size());
     hpx::util::high_resolution_timer t1;
 
     hpx::partitioned_vector<T> in(size, dist_policy);
@@ -120,8 +119,8 @@ void inclusive_scan_algo_tests_inplace_with_policy(
     double e1 = t1.elapsed();
     t1.restart();
 
-    hpx::parallel::inclusive_scan(policy,
-        in.begin(), in.end(), in.begin(), opt<T>(), val);
+    hpx::parallel::inclusive_scan(
+        policy, in.begin(), in.end(), in.begin(), opt<T>(), val);
 
     double e2 = t1.elapsed();
     t1.restart();
@@ -129,20 +128,20 @@ void inclusive_scan_algo_tests_inplace_with_policy(
     verify_values(in, ver);
 
     double e3 = t1.elapsed();
-    std::cout << std::setprecision(4) << "\t" << e1 << " " << e2 << " " << e3 << "\n";
+    std::cout << std::setprecision(4) << "\t" << e1 << " " << e2 << " " << e3
+              << "\n";
 }
 
 ///////////////////////////////////////////////////////////////////////////////
 
 template <typename T, typename DistPolicy, typename ExPolicy>
-void inclusive_scan_algo_tests_with_policy_async(
-    std::size_t size, DistPolicy const& dist_policy,
-    hpx::partitioned_vector<T>& in,
+void inclusive_scan_algo_tests_with_policy_async(std::size_t size,
+    DistPolicy const& dist_policy, hpx::partitioned_vector<T>& in,
     std::vector<T> ver, ExPolicy const& policy)
 {
     msg7(typeid(ExPolicy).name(), typeid(DistPolicy).name(), typeid(T).name(),
-        async, size,
-        dist_policy.get_num_partitions(), dist_policy.get_localities().size());
+        async, size, dist_policy.get_num_partitions(),
+        dist_policy.get_localities().size());
     hpx::util::high_resolution_timer t1;
 
     std::vector<T> out(in.size());
@@ -151,9 +150,8 @@ void inclusive_scan_algo_tests_with_policy_async(
     double e1 = t1.elapsed();
     t1.restart();
 
-    auto res =
-        hpx::parallel::inclusive_scan(policy,
-        in.begin(), in.end(), out.begin(), opt<T>(), val);
+    auto res = hpx::parallel::inclusive_scan(
+        policy, in.begin(), in.end(), out.begin(), opt<T>(), val);
     res.get();
 
     double e2 = t1.elapsed();
@@ -162,20 +160,21 @@ void inclusive_scan_algo_tests_with_policy_async(
     HPX_TEST(std::equal(out.begin(), out.end(), ver.begin()));
 
     double e3 = t1.elapsed();
-    std::cout << std::setprecision(4) << "\t" << e1 << " " << e2 << " " << e3 << "\n";
+    std::cout << std::setprecision(4) << "\t" << e1 << " " << e2 << " " << e3
+              << "\n";
 }
 
 template <typename T, typename DistPolicy, typename ExPolicy>
-void inclusive_scan_algo_tests_segmented_out_with_policy_async(
-    std::size_t size, DistPolicy const& in_dist_policy,
-    DistPolicy const& out_dist_policy,
+void inclusive_scan_algo_tests_segmented_out_with_policy_async(std::size_t size,
+    DistPolicy const& in_dist_policy, DistPolicy const& out_dist_policy,
     hpx::partitioned_vector<T>& in, hpx::partitioned_vector<T> out,
     std::vector<T> ver, ExPolicy const& policy)
 {
     msg9(typeid(ExPolicy).name(), typeid(DistPolicy).name(), typeid(T).name(),
-        async_segmented, size,
-        in_dist_policy.get_num_partitions(), in_dist_policy.get_localities().size(),
-        out_dist_policy.get_num_partitions(), out_dist_policy.get_localities().size());
+        async_segmented, size, in_dist_policy.get_num_partitions(),
+        in_dist_policy.get_localities().size(),
+        out_dist_policy.get_num_partitions(),
+        out_dist_policy.get_localities().size());
     hpx::util::high_resolution_timer t1;
 
     T val(0);
@@ -183,9 +182,8 @@ void inclusive_scan_algo_tests_segmented_out_with_policy_async(
     double e1 = t1.elapsed();
     t1.restart();
 
-    auto res =
-        hpx::parallel::inclusive_scan(policy,
-        in.begin(), in.end(), out.begin(), opt<T>(), val);
+    auto res = hpx::parallel::inclusive_scan(
+        policy, in.begin(), in.end(), out.begin(), opt<T>(), val);
     res.get();
 
     double e2 = t1.elapsed();
@@ -194,17 +192,17 @@ void inclusive_scan_algo_tests_segmented_out_with_policy_async(
     verify_values(out, ver);
 
     double e3 = t1.elapsed();
-    std::cout << std::setprecision(4) << "\t" << e1 << " " << e2 << " " << e3 << "\n";
+    std::cout << std::setprecision(4) << "\t" << e1 << " " << e2 << " " << e3
+              << "\n";
 }
 
 template <typename T, typename DistPolicy, typename ExPolicy>
-void inclusive_scan_algo_tests_inplace_with_policy_async(
-    std::size_t size, DistPolicy const& dist_policy,
-    std::vector<T> ver, ExPolicy const& policy)
+void inclusive_scan_algo_tests_inplace_with_policy_async(std::size_t size,
+    DistPolicy const& dist_policy, std::vector<T> ver, ExPolicy const& policy)
 {
     msg7(typeid(ExPolicy).name(), typeid(DistPolicy).name(), typeid(T).name(),
-        async_inplace, size,
-        dist_policy.get_num_partitions(), dist_policy.get_localities().size());
+        async_inplace, size, dist_policy.get_num_partitions(),
+        dist_policy.get_localities().size());
     hpx::util::high_resolution_timer t1;
 
     hpx::partitioned_vector<T> in(size, dist_policy);
@@ -215,9 +213,8 @@ void inclusive_scan_algo_tests_inplace_with_policy_async(
     double e1 = t1.elapsed();
     t1.restart();
 
-    auto res =
-        hpx::parallel::inclusive_scan(policy,
-        in.begin(), in.end(), in.begin(), opt<T>(), val);
+    auto res = hpx::parallel::inclusive_scan(
+        policy, in.begin(), in.end(), in.begin(), opt<T>(), val);
     res.get();
 
     double e2 = t1.elapsed();
@@ -226,9 +223,9 @@ void inclusive_scan_algo_tests_inplace_with_policy_async(
     verify_values(in, ver);
 
     double e3 = t1.elapsed();
-    std::cout << std::setprecision(4) << "\t" << e1 << " " << e2 << " " << e3 << "\n";
+    std::cout << std::setprecision(4) << "\t" << e1 << " " << e2 << " " << e3
+              << "\n";
 }
-
 
 ///////////////////////////////////////////////////////////////////////////////
 
@@ -250,10 +247,8 @@ void inclusive_scan_tests_with_policy(std::size_t size,
         ver.begin(), ver.end(), ver.begin(), val, opt<T>());
 
     //sync
-    inclusive_scan_algo_tests_with_policy<T>(
-        size, policy, in, ver, seq);
-    inclusive_scan_algo_tests_with_policy<T>(
-        size, policy, in, ver, par);
+    inclusive_scan_algo_tests_with_policy<T>(size, policy, in, ver, seq);
+    inclusive_scan_algo_tests_with_policy<T>(size, policy, in, ver, par);
 
     //async
     inclusive_scan_algo_tests_with_policy_async<T>(
@@ -263,8 +258,8 @@ void inclusive_scan_tests_with_policy(std::size_t size,
 }
 
 template <typename T, typename DistPolicy>
-void inclusive_scan_tests_segmented_out_with_policy(std::size_t size,
-    DistPolicy const& in_policy, DistPolicy const& out_policy)
+void inclusive_scan_tests_segmented_out_with_policy(
+    std::size_t size, DistPolicy const& in_policy, DistPolicy const& out_policy)
 {
     using namespace hpx::parallel::execution;
 
@@ -309,10 +304,8 @@ void inclusive_scan_tests_inplace_with_policy(std::size_t size,
         ver.begin(), ver.end(), ver.begin(), val, opt<T>());
 
     // sync
-    inclusive_scan_algo_tests_inplace_with_policy<T>(
-        size, policy, ver, seq);
-    inclusive_scan_algo_tests_inplace_with_policy<T>(
-        size, policy, ver, par);
+    inclusive_scan_algo_tests_inplace_with_policy<T>(size, policy, ver, seq);
+    inclusive_scan_algo_tests_inplace_with_policy<T>(size, policy, ver, par);
 
     // async
     inclusive_scan_algo_tests_inplace_with_policy_async<T>(
@@ -324,14 +317,16 @@ void inclusive_scan_tests_inplace_with_policy(std::size_t size,
 ///////////////////////////////////////////////////////////////////////////////
 
 template <typename T>
-void inclusive_scan_tests(std::vector<hpx::id_type> &localities)
+void inclusive_scan_tests(std::vector<hpx::id_type>& localities)
 {
     std::size_t const length = 1000000;
 
     inclusive_scan_tests_with_policy<T>(length, hpx::container_layout);
     inclusive_scan_tests_with_policy<T>(length, hpx::container_layout(3));
-    inclusive_scan_tests_with_policy<T>(length, hpx::container_layout(3, localities));
-    inclusive_scan_tests_with_policy<T>(length, hpx::container_layout(localities));
+    inclusive_scan_tests_with_policy<T>(
+        length, hpx::container_layout(3, localities));
+    inclusive_scan_tests_with_policy<T>(
+        length, hpx::container_layout(localities));
 
     inclusive_scan_tests_with_policy<T>(1000, hpx::container_layout(1000));
 
@@ -339,15 +334,14 @@ void inclusive_scan_tests(std::vector<hpx::id_type> &localities)
     inclusive_scan_tests_segmented_out_with_policy<T>(length,
         hpx::container_layout(localities), hpx::container_layout(localities));
 
-    inclusive_scan_tests_segmented_out_with_policy<T>(length,
-        hpx::container_layout(localities), hpx::container_layout(3));
+    inclusive_scan_tests_segmented_out_with_policy<T>(
+        length, hpx::container_layout(localities), hpx::container_layout(3));
 
-    inclusive_scan_tests_segmented_out_with_policy<T>(length,
-        hpx::container_layout(localities), hpx::container_layout(10));
+    inclusive_scan_tests_segmented_out_with_policy<T>(
+        length, hpx::container_layout(localities), hpx::container_layout(10));
 
-    inclusive_scan_tests_inplace_with_policy<T>(length,
-        hpx::container_layout(localities));
-
+    inclusive_scan_tests_inplace_with_policy<T>(
+        length, hpx::container_layout(localities));
 }
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -359,4 +353,3 @@ int main()
 
     return hpx::util::report_errors();
 }
-
