+++ tests/unit/parallel/segmented_algorithms/partitioned_vector_transform.cpp
@@ -4,7 +4,7 @@
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include <hpx/hpx_main.hpp>
-#include <hpx/include/partitioned_vector.hpp>
+#include <hpx/include/partitioned_vector_predef.hpp>
 #include <hpx/include/parallel_transform.hpp>
 #include <hpx/include/parallel_count.hpp>
 
@@ -12,10 +12,11 @@
 
 #include <cstddef>
 #include <vector>
+
 ///////////////////////////////////////////////////////////////////////////////
-// Define the vector types to be used.
-HPX_REGISTER_PARTITIONED_VECTOR(double);
-HPX_REGISTER_PARTITIONED_VECTOR(int);
+// The vector types to be used are defined in partitioned_vector module.
+// HPX_REGISTER_PARTITIONED_VECTOR(double);
+// HPX_REGISTER_PARTITIONED_VECTOR(int);
 
 template <typename U>
 struct pfo
@@ -26,6 +27,7 @@ struct pfo
         return U((T(2)*val));
     }
 };
+
 template <typename U>
 struct add
 {
@@ -101,38 +103,6 @@ void test_transform(ExPolicy && policy, hpx::partitioned_vector<T>& v,
     verify_values_count(policy, w, 2*val);
 }
 
-template <typename ExPolicy, typename T, typename U=T, typename V=T>
-void test_transform_binary(ExPolicy && policy, hpx::partitioned_vector<T>& v,
-    hpx::partitioned_vector<U>& w, hpx::partitioned_vector<V>& x, V val)
-{
-    verify_values(policy, v, val);
-    verify_values_count(policy, v, val);
-    verify_values(policy, w, val);
-    verify_values_count(policy, w, val);
-
-    hpx::parallel::transform(policy, v.begin(), v.end(), w.begin(),
-        x.begin(), add<V>());
-
-    verify_values(policy, x, 2*val);
-    verify_values_count(policy, x, 2*val);
-}
-
-template <typename ExPolicy, typename T, typename U=T, typename V=T>
-void test_transform_binary2(ExPolicy && policy, hpx::partitioned_vector<T>& v,
-    hpx::partitioned_vector<U>& w, hpx::partitioned_vector<V>& x, V val)
-{
-    verify_values(policy, v, val);
-    verify_values_count(policy, v, val);
-    verify_values(policy, w, val);
-    verify_values_count(policy, w, val);
-
-    hpx::parallel::transform(policy, v.begin(), v.end(), w.begin(), w.end(),
-        x.begin(), add<V>());
-
-    verify_values(policy, x, 2*val);
-    verify_values_count(policy, x, 2*val);
-}
-
 template <typename ExPolicy, typename T, typename U=T>
 void verify_values_count_async(ExPolicy && policy,
     hpx::partitioned_vector<T> const& v, U const& val)
@@ -159,40 +129,6 @@ void test_transform_async(ExPolicy && policy, hpx::partitioned_vector<T>& v,
 
     verify_values(policy, w, 2*val);
     verify_values_count_async(policy, w, 2*val);
-}
-
-template <typename ExPolicy, typename T, typename U=T, typename V=T>
-void test_transform_binary_async(ExPolicy && policy,
-    hpx::partitioned_vector<T>& v, hpx::partitioned_vector<U>& w,
-    hpx::partitioned_vector<V>& x, V val)
-{
-    verify_values(policy, v, val);
-    verify_values_count_async(policy, v, val);
-    verify_values(policy, w, val);
-    verify_values_count_async(policy, w, val);
-
-    hpx::parallel::transform(policy, v.begin(), v.end(), w.begin(),
-        x.begin(), add<V>()).get();
-
-    verify_values(policy, x, 2*val);
-    verify_values_count_async(policy, x, 2*val);
-}
-
-template <typename ExPolicy, typename T, typename U=T, typename V=T>
-void test_transform_binary2_async(ExPolicy && policy,
-    hpx::partitioned_vector<T>& v, hpx::partitioned_vector<U>& w,
-    hpx::partitioned_vector<V>& x, V val)
-{
-    verify_values(policy, v, val);
-    verify_values_count_async(policy, v, val);
-    verify_values(policy, w, val);
-    verify_values_count_async(policy, w, val);
-
-    hpx::parallel::transform(policy, v.begin(), v.end(), w.begin(), w.end(),
-        x.begin(), add<V>()).get();
-
-    verify_values(policy, x, 2*val);
-    verify_values_count_async(policy, x, 2*val);
 }
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -229,87 +165,14 @@ void transform_tests(std::vector<hpx::id_type> &localities)
             v, w, U(1));
     }
 }
-template <typename T, typename U=T, typename V=T>
-void transform_binary_tests(std::vector<hpx::id_type> &localities)
-{
-    std::size_t const length = 12;
-
-    {
-        hpx::partitioned_vector<T> v;
-        hpx::partitioned_vector<U> w;
-        hpx::partitioned_vector<V> x;
-        hpx::parallel::transform(hpx::parallel::execution::seq,
-            v.begin(), v.end(), w.begin(), x.begin(), add<V>());
-        hpx::parallel::transform(hpx::parallel::execution::par,
-            v.begin(), v.end(), w.begin(), x.begin(), add<V>());
-        hpx::parallel::transform(
-            hpx::parallel::execution::seq(hpx::parallel::execution::task),
-            v.begin(), v.end(), w.begin(), x.begin(), add<V>()).get();
-        hpx::parallel::transform(
-            hpx::parallel::execution::par(hpx::parallel::execution::task),
-            v.begin(), v.end(), w.begin(), x.begin(), add<V>()).get();
-    }
-
-    {
-        hpx::partitioned_vector<T> v(length, T(1),hpx::container_layout(localities));
-        hpx::partitioned_vector<U> w(length, U(1),hpx::container_layout(localities));
-        hpx::partitioned_vector<V> x(length,hpx::container_layout(localities));
-        test_transform_binary(hpx::parallel::execution::seq, v, w, x, V(1));
-        test_transform_binary(hpx::parallel::execution::par, v, w, x, V(1));
-        test_transform_binary_async(
-            hpx::parallel::execution::seq(hpx::parallel::execution::task),
-            v, w, x, V(1));
-        test_transform_binary_async(
-            hpx::parallel::execution::par(hpx::parallel::execution::task),
-            v, w, x, V(1));
-    }
-}
-
-template <typename T, typename U=T, typename V=T>
-void transform_binary2_tests(std::vector<hpx::id_type> &localities)
-{
-    std::size_t const length = 12;
-    {
-        hpx::partitioned_vector<T> v;
-        hpx::partitioned_vector<U> w;
-        hpx::partitioned_vector<V> x;
-        hpx::parallel::transform(hpx::parallel::execution::seq,
-            v.begin(), v.end(), w.begin(), w.end(), x.begin(), add<V>());
-        hpx::parallel::transform(hpx::parallel::execution::par,
-            v.begin(), v.end(), w.begin(), w.end(), x.begin(), add<V>());
-        hpx::parallel::transform(
-            hpx::parallel::execution::seq(hpx::parallel::execution::task),
-            v.begin(), v.end(), w.begin(), w.end(), x.begin(), add<V>()).get();
-        hpx::parallel::transform(
-            hpx::parallel::execution::par(hpx::parallel::execution::task),
-            v.begin(), v.end(), w.begin(), w.end(), x.begin(), add<V>()).get();
-    }
-
-    {
-        hpx::partitioned_vector<T> v(length, T(1),hpx::container_layout(localities));
-        hpx::partitioned_vector<U> w(length, U(1),hpx::container_layout(localities));
-        hpx::partitioned_vector<V> x(length,hpx::container_layout(localities));
-        test_transform_binary2(hpx::parallel::execution::seq, v, w, x, V(1));
-        test_transform_binary2(hpx::parallel::execution::par, v, w, x, V(1));
-        test_transform_binary2_async(
-            hpx::parallel::execution::seq(hpx::parallel::execution::task),
-            v, w, x, V(1));
-        test_transform_binary2_async(
-            hpx::parallel::execution::par(hpx::parallel::execution::task),
-            v, w, x, V(1));
-    }
-}
-///////////////////////////////////////////////////////////////////////////////
 
 ///////////////////////////////////////////////////////////////////////////////
 int main()
 {
     std::vector<hpx::id_type> localities = hpx::find_all_localities();
+
     transform_tests<int,int>(localities);
     transform_tests<double,int>(localities);
-    transform_binary_tests<int,int,double>(localities);
-    transform_binary_tests<double,double,int>(localities);
-    transform_binary2_tests<int,int,double>(localities);
-    transform_binary2_tests<double,double,int>(localities);
+
     return hpx::util::report_errors();
 }
