+++ tests/unit/parallel/segmented_algorithms/partitioned_vector_transform_scan.cpp
@@ -4,20 +4,20 @@
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include <hpx/hpx_main.hpp>
-#include <hpx/include/partitioned_vector.hpp>
+#include <hpx/include/partitioned_vector_predef.hpp>
 #include <hpx/include/parallel_transform_scan.hpp>
 
 #include <hpx/util/lightweight_test.hpp>
 
 #include <cstddef>
 #include <vector>
-///////////////////////////////////////////////////////////////////////////////
-// Define the vector types to be used.
-HPX_REGISTER_PARTITIONED_VECTOR(double);
-HPX_REGISTER_PARTITIONED_VECTOR(int);
 
 ///////////////////////////////////////////////////////////////////////////////
+// The vector types to be used are defined in partitioned_vector module.
+// HPX_REGISTER_PARTITIONED_VECTOR(double);
+// HPX_REGISTER_PARTITIONED_VECTOR(int);
 
+///////////////////////////////////////////////////////////////////////////////
 struct conv
 {
     template<typename T>
